mol-engine/requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
httpx>=0.28.0
pydantic>=2.0.0
asyncpg>=0.30.0
grpcio>=1.70.0
protobuf>=5.29.0
pyyaml>=6.0.0
